Some refactors and updates from lint issues
