+++ app/src/androidTest/java/net/trustly/trustlysdkdemoandroid/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package net.trustly.paywithmybanksdkdemoandroid
+package net.trustly.trustlysdkdemoandroid
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("net.trustly.paywithmybanksdkdemoandroid", appContext.packageName)
+        assertEquals("net.trustly.trustlysdkdemoandroid", appContext.packageName)
     }
 }
