+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PayWithMyBankSDKDemoAndroid"
         android:usesCleartextTraffic="true"
