+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">PayWithMyBankSDKDemoAndroid</string>
+    <string name="app_name">TrustlySDKDemoAndroid</string>
     <string name="connect_my_bank">Connect my bank</string>
     <string name="success">Success!</string>
     <string name="canceled">Canceled =(</string>
